one/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
chardet==3.0.4
Django==5.0.6
django-cors-headers==4.4.0
django-extensions==3.2.3
django-jazzmin==3.0.0
django-modeltranslation==0.19.4
django-sslserver==0.22
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.6.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
inflection==0.5.1
packaging==24.1
pillow==10.3.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
transliterate==1.10.2
typing_extensions==4.12.2
uritemplate==4.1.1
